requirements.txt
tensorflow
pillow
matplotlib
gradio